Updated handler.py
This fixes an issue which occurs when a user can specify a list of URLs of which some of them may be dead. This often occurs when a bug bounty hunter discovers old domains or links and puts them through their tool-set. 

This stops Konan from dying if a timeout occurs and will continue through the list.